changes to waterfox install script for updates and install repairs.

+++ installers/packages/waterfox/fedora_install.sh
@@ -9,6 +9,13 @@ fi
 
 # Variables
 WATERFOX_VERSION="G4.0.7"
+BASE_PATH=$(dirname $0)
+
+# If in directory script is located.
+if [ -z "$BASE_PATH" ]
+then
+    BASE_PATH="."
+fi
 
 # Create required directories if they don't exist
 mkdir -p /tmp /opt
@@ -21,11 +28,15 @@ curl -L "https://github.com/WaterfoxCo/Waterfox/releases/download/${WATERFOX_VER
 
 # Install Waterfox
 tar -xf /tmp/waterfox.tar.bz2 -C /tmp
+if [ -d /opt/waterfox ]; then rm -rf /opt/waterfox; fi
 mv /tmp/waterfox /opt/waterfox
 chmod +x /opt/waterfox/waterfox
-ln -s /opt/waterfox/waterfox /usr/bin/waterfox
-cp waterfox.desktop /opt/waterfox/waterfox.desktop
-ln -s /opt/waterfox/waterfox.desktop /usr/local/share/applications/waterfox.desktop
+if [ ! -f /usr/bin/waterfox ]; then ln -s /opt/waterfox/waterfox /usr/bin/waterfox; fi
+cp $BASE_PATH/waterfox.desktop /opt/waterfox/waterfox.desktop
+if [ ! -f /usr/local/share/applications/waterfox.desktop ]
+then
+    ln -s /opt/waterfox/waterfox.desktop /usr/local/share/applications/waterfox.desktop
+fi
 
 # Installation Check
 WATERFOX_PATH=$(which waterfox)
